+++ library/docs/changelog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 - Using `DeprecationLevel.Hidden` is temporarily treated as an error until the [compiler issue](https://github.com/MiSikora/laboratory/issues/62) is fixed.
 
+### Fixed
+- Make updates to the in-memory `Laboratory` atomic.
+
 ## [0.9.5] - 2020-12-03
 
 ### Added