+++ library/docs/changelog.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Upgrade Kotlin to `1.4.21`
+
 ### Deprecated
 - Using `DeprecationLevel.Hidden` is temporarily treated as an error until the [compiler issue](https://github.com/MiSikora/laboratory/issues/62) is fixed.