+++ library/docs/changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+- Use description as feature flag's KDoc content
 - Upgrade Kotlin to `1.4.21`
 
 ### Deprecated