+++ library/docs/changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.6] - 2020-12-15
+
 ### Added
 - `FeatureFactory` can be now appended to another with plus operator.
 - `DefaultOptionFactory` can be now appended to another with plus operator.
@@ -246,7 +248,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/MiSikora/laboratory/compare/0.9.5...HEAD
+[Unreleased]: https://github.com/MiSikora/laboratory/compare/0.9.6...HEAD
+[0.9.6]: https://github.com/MiSikora/laboratory/releases/tag/0.9.6
 [0.9.5]: https://github.com/MiSikora/laboratory/releases/tag/0.9.5
 [0.9.4]: https://github.com/MiSikora/laboratory/releases/tag/0.9.4
 [0.9.3]: https://github.com/MiSikora/laboratory/releases/tag/0.9.3